+++ Messaging/Plugin.Messaging.Android.Sample/Resources/Resource.Designer.cs
@@ -63,26 +63,29 @@ namespace Plugin.Messaging.Sample.Android
 			// aapt resource value: 0x7f060000
 			public const int ButtonMakePhoneCall = 2131099648;
 			
-			// aapt resource value: 0x7f060006
-			public const int ButtonSendAttachmentEmail = 2131099654;
+			// aapt resource value: 0x7f060001
+			public const int ButtonMakePhoneCallAutoDial = 2131099649;
 			
 			// aapt resource value: 0x7f060007
-			public const int ButtonSendAttachmentEmailPCL = 2131099655;
+			public const int ButtonSendAttachmentEmail = 2131099655;
 			
-			// aapt resource value: 0x7f060003
-			public const int ButtonSendEmail = 2131099651;
+			// aapt resource value: 0x7f060008
+			public const int ButtonSendAttachmentEmailPCL = 2131099656;
 			
 			// aapt resource value: 0x7f060004
-			public const int ButtonSendEmailStrictMode = 2131099652;
+			public const int ButtonSendEmail = 2131099652;
 			
 			// aapt resource value: 0x7f060005
-			public const int ButtonSendHtmlEmail = 2131099653;
+			public const int ButtonSendEmailStrictMode = 2131099653;
+			
+			// aapt resource value: 0x7f060006
+			public const int ButtonSendHtmlEmail = 2131099654;
+			
+			// aapt resource value: 0x7f060003
+			public const int ButtonSendMultipleSms = 2131099651;
 			
 			// aapt resource value: 0x7f060002
-			public const int ButtonSendMultipleSms = 2131099650;
-			
-			// aapt resource value: 0x7f060001
-			public const int ButtonSendSms = 2131099649;
+			public const int ButtonSendSms = 2131099650;
 			
 			static Id()
 			{