+++ Messaging/ChangeLog.md
@@ -3,7 +3,9 @@
 ### [4.0.0] ###
 - Add support for sending SMS to multiple recipients
 - Android: Add support for using `FileProvider` to add attachments using content Uri's on Android Nougat and later
-- Android: Add new `EmailMessageBuilder.UseStrictMode` to filter list of apps to only email apps and not other text messaging or social apps. **Unfortunately adding attachments when using StrictMode does not seem to work, and is therefore currently not supported.**
+- Android: Add new `Settings` class to configure Android specific behavior when sending emails/making phone calls (see next bullets).  Access from Android project using `CrossMessaging.Current.Settings()` extension method.
+- Android: Add new `EmailSettings.UseStrictMode` flag (default value `false`) to filter list of apps to only email apps and not other text messaging or social apps. **Unfortunately adding attachments when using StrictMode does not seem to work, and is therefore currently not supported.**
+- Android: Add new `PhoneSettings.AutoDial` flag (default value `false`) to automatically phone the number instead of only showing the phone dialer with the number populated. **Please note using this settings requires the `android.permission.CALL_PHONE` added to the manifest file.**
 - **Breaking Change**: Remove iOS Classic support
 - **Breaking Change**: Remove Windows Phone 8.0 and 8.1 support
 - **Breaking Change**: Reworked `EmailMessageBuilder.WithAttachment` platform API to provide consistent API